# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Hospital Management System.",
        "title": "HMS API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/ekastn",
            "email": "eka@voixes.studio"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5021",
    "basePath": "/api",
    "paths": {
...
        "/patients": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of all registered patients.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Get all patients",
                "responses": {
                    "200": {
                        "description": "List of patients",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.PatientDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve patients",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new patient entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Create a new patient",
                "parameters": [
                    {
                        "description": "Patient object to be created",
                        "name": "patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreatePatientRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Patient created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation failed",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create patient",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/patients/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a single patient by their ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Get patient by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.PatientDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Patient not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve patient",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update details of an existing patient.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Update an existing patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Patient object with updated fields",
                        "name": "patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdatePatientRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Patient updated successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation failed",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Patient not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update patient",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a patient entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Delete a patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Patient deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete patient",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/patients/{id}/detail": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve comprehensive information for a single patient, including recent appointments and medical history.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Get detailed patient information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient details with appointments and medical history",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.PatientDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Patient not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve patient details",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.AppointmentDTO": {
            "type": "object",
            "required": [
                "dateTime",
                "doctorId",
                "duration",
                "location",
                "patientId",
                "status",
                "type"
            ],
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2025-07-17T09:00:00Z"
                },
                "dateTime": {
                    "type": "string",
                    "example": "2025-07-17T10:00:00Z"
                },
                "doctorId": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000003"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000001"
                },
                "location": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "Room 101"
                },
                "notes": {
                    "type": "string",
                    "maxLength": 500,
                    "example": "Patient complained of headache"
                },
                "patientHistory": {
                    "type": "string",
                    "maxLength": 1000,
                    "example": "No significant medical history"
                },
                "patientId": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000002"
                },
                "status": {
                    "enum": [
                        "Scheduled",
                        "Confirmed",
                        "Completed",
                        "Cancelled"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.AppointmentStatus"
                        }
                    ],
                    "example": "Scheduled"
                },
                "type": {
                    "enum": [
                        "check-up",
                        "follow-up",
                        "consultation",
                        "procedure",
                        "emergency"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.AppointmentType"
                        }
                    ],
                    "example": "check-up"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2025-07-17T09:00:00Z"
                }
            }
        },
...
        "domain.AppointmentStatus": {
            "type": "string",
            "enum": [
                "Scheduled",
                "Confirmed",
                "Completed",
                "Cancelled"
            ],
            "x-enum-varnames": [
                "AppointmentStatusScheduled",
                "AppointmentStatusConfirmed",
                "AppointmentStatusCompleted",
                "AppointmentStatusCancelled"
            ]
        },
        "domain.AppointmentType": {
            "type": "string",
            "enum": [
                "check-up",
                "follow-up",
                "consultation",
                "procedure",
                "emergency"
            ],
            "x-enum-varnames": [
                "AppointmentTypeCheckUp",
                "AppointmentTypeFollowUp",
                "AppointmentTypeConsultation",
                "AppointmentTypeProcedure",
                "AppointmentTypeEmergency"
            ]
        },
...
        "domain.CreatePatientRequest": {
            "description": "Request body for creating a new patient",
            "type": "object",
            "required": [
                "address",
                "age",
                "email",
                "gender",
                "name",
                "phone"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 10,
                    "example": "123 Main St, Anytown, USA"
                },
                "age": {
                    "type": "integer",
                    "maximum": 120
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "Male",
                        "Female",
                        "Other"
                    ],
                    "example": "Male"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "John Doe"
                },
                "phone": {
                    "type": "string",
                    "example": "+1234567890"
                }
            }
        },
...
        "domain.MedicalRecordDTO": {
            "description": "Medical record data transfer object",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2025-07-17T09:00:00Z"
                },
                "date": {
                    "type": "string",
                    "example": "2025-07-17T10:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Patient presented with flu-like symptoms."
                },
                "diagnosis": {
                    "type": "string",
                    "example": "Influenza A"
                },
                "doctorId": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000003"
                },
                "id": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000001"
                },
                "notes": {
                    "type": "string",
                    "example": "Advised patient to stay hydrated."
                },
                "patientId": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000002"
                },
                "recordType": {
                    "type": "string",
                    "example": "checkup"
                },
                "treatment": {
                    "type": "string",
                    "example": "Prescribed Tamiflu and rest."
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2025-07-17T09:00:00Z"
                }
            }
        },
        "domain.PatientDTO": {
            "description": "Patient data transfer object",
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "123 Main St"
                },
                "age": {
                    "type": "integer"
                },
                "email": {
                    "type": "string",
                    "example": "jane.doe@example.com"
                },
                "gender": {
                    "type": "string",
                    "example": "Female"
                },
                "id": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000001"
                },
                "lastVisit": {
                    "type": "string",
                    "example": "2025-07-17T10:00:00Z"
                },
                "name": {
                    "type": "string",
                    "example": "Jane Doe"
                },
                "phone": {
                    "type": "string",
                    "example": "1234567890"
                }
            }
        },
        "domain.PatientDetailResponse": {
            "description": "Detailed patient information including recent appointments and medical history",
            "type": "object",
            "properties": {
                "medicalHistory": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.MedicalRecordDTO"
                    }
                },
                "patient": {
                    "$ref": "#/definitions/domain.PatientDTO"
                },
                "recentAppointments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.AppointmentDTO"
                    }
                }
            }
        },
...
        "domain.UpdatePatientRequest": {
            "description": "Request body for updating an existing patient",
            "type": "object",
            "required": [
                "address",
                "age",
                "email",
                "gender",
                "name",
                "phone"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 10,
                    "example": "456 Oak Ave, Somewhere, USA"
                },
                "age": {
                    "type": "integer",
                    "maximum": 120
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "Male",
                        "Female",
                        "Other"
                    ],
                    "example": "Male"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "John Doe Jr."
                },
                "phone": {
                    "type": "string",
                    "example": "+1234567890"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Error message"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "utils.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}